Primitive styles for the project navigation (not DRY)
Unfortunately with comments and notifications using independent
controllers we cannot use nav-tabs to swap between views. I'll have to
create a better alternative than I have currently, but this is the
first primitive set up. The edit view allows us to navigate to every
section, while the comment and notification index pages do not--I will
add this feature as soon as I decide on something that makes long-term
sense and is DRY.