Project tiles update
Made each tile a link to the show page for the corresponding project.
Updated the links in each tile: 1. details (links to show page) 2.
comments (links to comments) 3. like (this is hidden for now as it's
not necessary at the moment). Finally, made a small change to the side
menu -- the 'notifications' link is hidden unless the project has
outstanding notifications